% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtrair.R
\name{subitrair}
\alias{subitrair}
\title{Função para subtrair}
\usage{
subitrair(a, b)
}
\arguments{
\item{a}{um número}

\item{b}{outro número}
}
\description{
Função para subtrair
}
\examples{
subtrair(3,4)

}
